+++ src/test/cases/routes/course.js
@@ -1,0 +1,12 @@
+const constructPortfolios = require('../../../main/routes/helper_functions');
+const { expect } = require('../../chai');
+
+describe('Route - course', () => {
+
+	describe('course: get /', () => {
+
+        
+
+	})
+
+})
